Remove duplicate folder
