+++ Sources/Fullscreen/TransactionView/TransactionAlertView/TransactionAlertViewModel.swift
@@ -6,6 +6,6 @@ import Foundation
 
 public protocol TransactionAlertViewModel {
     var title: String { get set }
-    var message: String { get set }
-    var imageUrl: String? { get set }
+    var body: String { get set }
+    var imageIdentifier: String { get set }
 }
